+++ src/cmem/markdown/Markdown.tsx
@@ -1,6 +1,9 @@
 import React from "react";
 import ReactMarkdown from "react-markdown";
 import { PluggableList } from "react-markdown/lib/react-markdown";
+import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
+import { materialLight } from "react-syntax-highlighter/dist/esm/styles/prism";
+
 //@ts-ignore
 import remarkTypograf from "@mavrin/remark-typograf";
 import rehypeRaw from "rehype-raw";
@@ -132,6 +135,25 @@ export const Markdown = ({
                   return linkTarget as React.HTMLAttributeAnchorTarget;
               }
             : undefined,
+        components: {
+            code(props: any) {
+                const { children, className, node, ...rest } = props;
+                const match = /language-(\w+)/.exec(className || "");
+                return match ? (
+                    <SyntaxHighlighter
+                        {...rest}
+                        PreTag="div"
+                        children={String(children).replace(/\n$/, "")}
+                        language={match[1]}
+                        style={materialLight}
+                    />
+                ) : (
+                    <code {...rest} className={className}>
+                        {children}
+                    </code>
+                );
+            },
+        },
     };
     allowedElements && (reactMarkdownProperties.allowedElements = allowedElements);
     reHypePlugins &&