+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+-   `<Markdown />`
+    -   Do syntax highlighting when a class name is set in the form `language-<LANGUAGE_NAME>`.
+
+### Fixed
+
+-   `<MultiSuggestField />`
+    -   Added logic for contolled state for the component if `selectedItems` and `onSelection` passed.
+
 ## [23.6.0] - 2024-04-17
 
 ### Added